+++ src/main/kotlin/com/github/srwi/pixellens/actions/ViewAsImageAction.kt
@@ -9,29 +9,35 @@ import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.xdebugger.impl.ui.tree.actions.XDebuggerTreeActionBase
 import com.jetbrains.python.debugger.PyDebugValue
+import javax.swing.SwingUtilities
 
 class ViewAsImageAction : AnAction() {
-
     override fun actionPerformed(e: AnActionEvent) {
         val project = e.project ?: return
-        try {
-            val value = XDebuggerTreeActionBase.getSelectedValue(e.dataContext) as PyDebugValue? ?: return
-            val imageProvider = ImageProviderFactory.getImageProvider(value.typeQualifier as String)
-            val batch = imageProvider.getDataByVariableName(value.frameAccessor, value.name)
+        val value = XDebuggerTreeActionBase.getSelectedValue(e.dataContext) as PyDebugValue? ?: return
+        val imageProvider = ImageProviderFactory.getImageProvider(value.typeQualifier as String)
+        val futureBatch = imageProvider.getDataByVariableName(project, value.frameAccessor, value.name)
+
+        futureBatch.thenAccept { batch ->
             batch.data.normalized = UserSettings.normalizeEnabled
             batch.data.channelsFirst = UserSettings.transposeEnabled
             batch.data.reversedChannels = UserSettings.reverseChannelsEnabled
             batch.data.grayscaleColormap = UserSettings.applyColormapEnabled
 
-            ImageViewer(project, batch).show()
-        } catch (ex: Exception) {
-            val formattedException = ex.message + "\n" + ex.stackTrace.joinToString("\n")
-            ErrorMessageDialog(
-                project,
-                "Error",
-                "The selected data could not be viewed as image.",
-                formattedException
-            ).show()
+            SwingUtilities.invokeLater {
+                ImageViewer(project, batch).show()
+            }
+        }.exceptionally { throwable ->
+            val formattedException = throwable.message + "\n" + throwable.stackTrace.joinToString("\n")
+            SwingUtilities.invokeLater {
+                ErrorMessageDialog(
+                    project,
+                    "Error",
+                    "The selected data could not be viewed as image.",
+                    formattedException
+                ).show()
+            }
+            null
         }
     }