+++ src/main/kotlin/com/github/srwi/pixellens/imageProviders/NumpyImageProvider.kt
@@ -1,43 +1,32 @@
 package com.github.srwi.pixellens.imageProviders
 
-import com.github.srwi.pixellens.interop.Python.evaluateExpression
-import com.github.srwi.pixellens.interop.Python.executeStatement
 import com.jetbrains.python.debugger.PyDebugValue
-import com.jetbrains.python.debugger.PyFrameAccessor
 
 class NumpyImageProvider : ImageProvider() {
-    override fun getPayload(frameAccessor: PyFrameAccessor, name: String): Payload {
-        val command = """
+    override fun getDataPreparationCommand(variableName: String): String {
+        return """
             import base64 as __tmp_base64
             import json as __tmp_json
             import numpy as __tmp_np
             
-            __tmp_img_array = $name
+            __tmp_img_array = $variableName
             __tmp_img_bytes = __tmp_np.ascontiguousarray(__tmp_img_array).tobytes()
-            __tmp_img_base64 = __tmp_base64.b64encode(__tmp_img_bytes).decode('utf-8')
-            __tmp_img_payload = {
-                'name': '$name',
-                'data': __tmp_img_base64,
-                'metadata': {
-                    'shape': __tmp_img_array.shape,
-                    'dtype': str(__tmp_img_array.dtype)
-                }
+            $Base64DataVariableName = __tmp_base64.b64encode(__tmp_img_bytes).decode('utf-8')
+            $MetadataVariableName = {
+                'name': '$variableName',
+                'length': len($Base64DataVariableName),
+                'shape': __tmp_img_array.shape,
+                'dtype': str(__tmp_img_array.dtype)
             }
-            __tmp_img_data = __tmp_json.dumps(__tmp_img_payload)
+            $MetadataVariableName = __tmp_json.dumps($MetadataVariableName)
         """.trimIndent()
+    }
 
-        val cleanupCommand = """
+    override fun getCleanupCommand(): String {
+        return """
             del __tmp_base64, __tmp_json, __tmp_np
-            del __tmp_img_array, __tmp_img_bytes, __tmp_img_base64, __tmp_img_payload, __tmp_img_data
+            del __tmp_img_array, __tmp_img_bytes, $MetadataVariableName, $Base64DataVariableName
         """.trimIndent()
-
-        try {
-            executeStatement(frameAccessor, command)
-            val data = evaluateExpression(frameAccessor, "__tmp_img_data")?.value
-            return deserializeJsonPayload(data ?: throw IllegalStateException("Failed to retrieve data"))
-        } finally {
-            executeStatement(frameAccessor, cleanupCommand)
-        }
     }
 
     override fun typeSupported(value: PyDebugValue): Boolean {