+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ## [Unreleased]
 
+- Added progress indicator while retrieving image data
+
+## [0.1.1] - 2024-08-05
+
+### Added
+
+- Added support for some missing PIL types
+
 ## [0.1.0] - 2024-08-04
 
 ### Added